further size_t fixes

+++ src/util/string_container.cpp
@@ -107,7 +107,7 @@ unsigned string_containert::get(const char *s)
   if(it!=hash_table.end())
     return it->second;
 
-  unsigned r=hash_table.size();
+  size_t r=hash_table.size();
 
   // these are stable
   string_list.push_back(std::string(s));
@@ -142,7 +142,7 @@ unsigned string_containert::get(const std::string &s)
   if(it!=hash_table.end())
     return it->second;
 
-  unsigned r=hash_table.size();
+  size_t r=hash_table.size();
 
   // these are stable
   string_list.push_back(s);
